feat: refactor data receiver
