+++ src/context.rs
@@ -2,8 +2,7 @@ use crate::utils::Bip32Path;
 
 pub struct TxContext {
     pub buffer_pos: usize,
-    pub path: Bip32Path,
-    pub review_finished: bool,
+    pub done: bool
 }
 
 // Implement constructor for TxInfo with default values
@@ -12,19 +11,13 @@ impl TxContext {
     pub fn new() -> TxContext {
         TxContext {
             buffer_pos: 0,
-            path: Default::default(),
-            review_finished: false,
+            done: false
         }
     }
-    // Get review status
-    #[allow(dead_code)]
-    pub fn finished(&self) -> bool {
-        self.review_finished
-    }
+
     // Implement reset for TxInfo
     pub fn reset(&mut self) {
         self.buffer_pos = 0;
-        self.path = Default::default();
-        self.review_finished = false;
+        self.done = false;
     }
 }
